Update README, folder structure

+++ scheduler_v1.py
@@ -4,6 +4,7 @@ import random
 from datetime import datetime, timedelta
 import logging
 from collections import deque, defaultdict
+import sys
 
 # Configure logging
 log_directory = os.path.join(os.path.dirname(__file__), "logging")
@@ -12,6 +13,7 @@ os.makedirs(log_directory, exist_ok=True)
 logging.basicConfig(
     level=logging.INFO,
     format="%(asctime)s - %(levelname)s - %(message)s",
+    datefmt="%m-%d-%Y %H:%M:%s",
     handlers=[
         # logging.StreamHandler(),  # Comment out to stop logging to console
         logging.FileHandler(
@@ -20,20 +22,80 @@ logging.basicConfig(
     ],
 )
 
+# Set the paths to the input and output files
+team_list_path = os.path.join("working", "team_list.xlsx")
+assignments_path = os.path.join("working", "assignments.xlsx")
+
+WEEKS_IN_YEAR = 52
+
+def read_employee_data(file_path):
+    """Read employee data from the Excel spreadsheet."""
+    try:
+        # df = pd.read_excel(file_path)
+        df = pd.read_excel(file_path)
+
+        # Clean column names by removing leading and trailing spaces
+        df.columns = df.columns.str.strip()
+
+        # Validate columns
+        expected_columns = ["Name", "Email", "Available"]
+        actual_columns_stripped = [col.strip() for col in df.columns]
+
+        if set(actual_columns_stripped) != set(expected_columns):
+            raise ValueError(
+                f"Expected Columns: {expected_columns}\nActual Columns: {actual_columns_stripped}"
+            )
+
+        return df
+
+    except Exception as e:
+        logging.error(f"Error reading employee data: {e}")
+        return None
+
+
+employee_data = read_employee_data(team_list_path)
 
 # Log the start of the script execution
 logging.info("Script execution started.")
 
 # Define the assignment data log file
-assignment_data_log_path = os.path.join(log_directory, "assignment_data_log.csv")
+assignment_data_log_filename = "assignment_data_log.csv"
+assignment_data_log_path = os.path.join(
+    os.path.dirname(__file__), "logging", assignment_data_log_filename
+)
+
+# Check if the assignment_data_log.csv file exists, create it if not
+if not os.path.exists(assignment_data_log_path):
+    with open(assignment_data_log_path, "w") as log_file:
+        log_file.write(
+            "Tech, Number of Assignments, Last Assignment Date, Workload History\n"
+        )
+
+
+# Load previous employee data and detect changes
+previous_employee_data = read_employee_data(team_list_path)
+
+
+def check_and_exit_if_no_changes(current_employee_data, previous_employee_data):
+    """Check for changes in employee availability and exit if no changes."""
+    if current_employee_data is None or previous_employee_data is None:
+        log_activity("Error reading employee data. Exiting script.")
+        logging.info("Script execution completed with errors.\n")
+        sys.exit(1)
+
+    if current_employee_data.equals(previous_employee_data):
+        log_activity("No changes in availability. Exiting script.")
+        logging.info("Script execution completed.\n")
+        sys.exit(0)
 
 
 def initialize_assignment_data_log():
-    # Initialize the assignment data log file
-    with open(assignment_data_log_path, "w") as log_file:
-        log_file.write(
-            "Tech,Number of Assignments,Last Assignment Date,Workload History\n"
-        )
+    # Check if the assignment_data_log.csv file exists
+    if not os.path.exists(assignment_data_log_path):
+        with open(assignment_data_log_path, "w") as log_file:
+            log_file.write(
+                "Tech, Number of Assignments, Last Assignment Date, Workload History\n"
+            )
 
 
 def load_assignment_history():
@@ -69,55 +131,52 @@ def load_assignment_history():
 def log_assignment_data(assignment_data):
     # Log assignment data to the assignment_data_log file
     with open(assignment_data_log_path, "a") as log_file:
+        # Write assignment data for each tech
         for tech, data in assignment_data.items():
-            log_file.write(
-                f"{tech},{data['num_assignments']},{data['last_assignment_date']},{','.join(data['workload_history'])}\n"
-            )
+            # Check if 'num_assignments' key exists in the data dictionary
+            if "num_assignments" in data:
+                log_file.write(
+                    f"{tech},{data['num_assignments']},{data['last_assignment_date']},{','.join(data['workload_history'])}\n"
+                )
+
+
+def log_activity(activity_description):
+    """Log activities to scheduler_events.log."""
+    formatted_timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+    logging.info(f"{formatted_timestamp} - {activity_description}")
 
 
 def backup_existing_assignments():
-    file_name = "assignments.xlsx"
+    file_name = "assignment_data_log.csv"
     history_directory = "history"
 
-    working_directory = "working"
-    file_path = os.path.join(working_directory, file_name)
+    file_path = os.path.join(log_directory, file_name)
 
     if os.path.exists(file_path):
         timestamp_str = datetime.now().strftime("%m-%d-%Y_%H-%M-%S")
         new_file_name = os.path.join(
-            history_directory, f"assignments_as_of_{timestamp_str}.xlsx"
+            history_directory, f"{file_name}_as_of_{timestamp_str}.csv"
         )
 
         os.rename(file_path, new_file_name)
-        logging.info("Assignment file backed up to {}".format(new_file_name))
-        # print(f"Existing {file_name} backed up to {new_file_name}")
+        logging.info("Assignment data log backed up to {}".format(new_file_name))
+
+        # Keep only the last 3 backups, delete older ones
+        delete_old_backups(history_directory, file_name, keep_latest=3)
     else:
-        logging.info("No existing assignments to back up.")
+        logging.info("No existing assignment data log to back up.")
 
 
-def read_employee_data(file_path):
-    """Read employee data from the Excel spreadsheet."""
-    try:
-        # df = pd.read_excel(file_path)
-        df = pd.read_excel(os.path.join("working", file_path))
+def delete_old_backups(directory, base_name, keep_latest):
+    files = [f for f in os.listdir(directory) if f.startswith(f"{base_name}_as_of_")]
+    files.sort(key=lambda x: os.path.getctime(os.path.join(directory, x)), reverse=True)
 
-        # Clean column names by removing leading and trailing spaces
-        df.columns = df.columns.str.strip()
-
-        # Validate columns
-        expected_columns = ["Name", "Email", "Available"]
-        actual_columns_stripped = [col.strip() for col in df.columns]
-
-        if set(actual_columns_stripped) != set(expected_columns):
-            raise ValueError(
-                f"Expected Columns: {expected_columns}\nActual Columns: {actual_columns_stripped}"
-            )
-
-        return df
-
-    except Exception as e:
-        logging.error(f"Error reading employee data: {e}")
-        return None
+    # Delete older backups, keep only the last 'keep_latest'
+    for old_file in files[keep_latest:]:
+        os.remove(os.path.join(directory, old_file))
+        logging.info(
+            f"Retain maximum of 3 saved logs. Old assignment data log backup deleted: {old_file}"
+        )
 
 
 def get_email_addresses(employee_data, pair):
@@ -181,6 +240,106 @@ def load_and_detect_changes(previous_data_path, current_data, log_changes=True):
 
     return previous_employee_data
 
+
+def calculate_week_dates(current_date, week):
+    """Calculate the start and end dates of a week."""
+    start_date = current_date + timedelta(
+        days=((week - 1) * 7) + (0 - current_date.weekday()) % 7
+    )
+    end_date = start_date + timedelta(days=4)
+    return start_date, end_date
+
+
+def generate_paired_employees(
+    employee_data, weights, assigned_pairs_queue, max_assignments
+):
+    """Generate a pair of employees for assignment."""
+    paired_employees = None
+    while True:
+        normalized_weights = normalize_weights(weights, max_assignments)
+        paired_employees = random.choices(
+            employee_data[employee_data["Available"] == "yes"]["Name"].tolist(),
+            weights=normalized_weights,
+            k=2,
+        )
+        if all(pair not in assigned_pairs_queue for pair in paired_employees):
+            break
+    assigned_pairs_queue.extend(paired_employees)
+    return paired_employees
+
+
+def update_weights(weights, history=None):
+    """Update weights based on certain criteria."""
+    return adjust_weights(weights, history)
+
+
+def adjust_weights(weights, history):
+    """Adjust weights based on certain criteria and historical data."""
+    adjusted_weights = []
+
+    for tech, weight in zip(employee_data["Name"], weights):
+        # Consider the historical number of assignments for each tech
+        num_assignments = history[tech]["num_assignments"]
+
+        # Modify the weight based on historical data (customize this part)
+        adjusted_weight = weight * (1.0 + 0.1 * num_assignments)  # Example adjustment
+
+        adjusted_weights.append(adjusted_weight)
+
+    return adjusted_weights
+
+
+def update_assignment_data(assignment_data, paired_employees, end_date, start_date):
+    """Update assignment data for each employee."""
+    for tech in paired_employees:
+        assignment_data[tech]["num_assignments"] += 1
+        assignment_data[tech]["last_assignment_date"] = end_date.strftime("%m-%d-%Y")
+        assignment_data[tech]["workload_history"].append(
+            start_date.strftime("%m-%d-%Y")
+        )
+
+
+def write_to_excel(schedule):
+    """Write the schedule to an Excel file with the specified format."""
+    flat_schedule = []
+
+    for week, data in schedule.items():
+        start_date = datetime.strptime(data["start_date"], "%m-%d-%Y")
+        end_date = datetime.strptime(data["end_date"], "%m-%d-%Y")
+
+        flat_schedule.append(
+            {
+                "start_date": start_date.strftime("%m-%d-%Y"),
+                "end_date": end_date.strftime("%m-%d-%Y"),
+                "Agent 1": data["pair"][0],
+                "Email1": data["email_addresses"][0],
+                "Agent 2": data["pair"][1],
+                "Email2": data["email_addresses"][1],
+            }
+        )
+
+    df = pd.DataFrame(flat_schedule)
+
+    # Specify the path within the "working" directory
+    working_directory = "working"
+    output_file_path = os.path.join(working_directory, "assignments.xlsx")
+
+    # Write to Excel file with the specified columns
+    df.to_excel(output_file_path, index=False)
+
+    # Logging to confirm the file creation
+    if not os.path.exists(output_file_path):
+        logging.warning(
+            f"Assignments file not found at {output_file_path}. Creating a new file."
+        )
+        with open(output_file_path, "w") as new_file:
+            new_file.write("Start Date, End Date, Agent 1, Email1, Agent 2, Email2\n")
+
+        logging.info(f"Assignments file created at {output_file_path}")
+    else:
+        logging.info(f"Assignments written to {output_file_path}")
+
+
 def generate_rotation_schedule(employee_data, weeks_in_year):
     schedule = {}
 
@@ -230,116 +389,27 @@ def generate_rotation_schedule(employee_data, weeks_in_year):
 
     return schedule
 
-
-
-def calculate_week_dates(current_date, week):
-    """Calculate the start and end dates of a week."""
-    start_date = current_date + timedelta(
-        days=((week - 1) * 7) + (0 - current_date.weekday()) % 7
-    )
-    end_date = start_date + timedelta(days=4)
-    return start_date, end_date
-
-
-def generate_paired_employees(
-    employee_data, weights, assigned_pairs_queue, max_assignments
-):
-    """Generate a pair of employees for assignment."""
-    paired_employees = None
-    while True:
-        normalized_weights = normalize_weights(weights, max_assignments)
-        paired_employees = random.choices(
-            employee_data[employee_data["Available"] == "yes"]["Name"].tolist(),
-            weights=normalized_weights,
-            k=2,
-        )
-        if all(pair not in assigned_pairs_queue for pair in paired_employees):
-            break
-    assigned_pairs_queue.extend(paired_employees)
-    return paired_employees
-
-
-def update_weights(weights, history=None):
-    """Update weights based on certain criteria."""
-    return adjust_weights(weights, history)
-
-def adjust_weights(weights, history):
-    """Adjust weights based on certain criteria and historical data."""
-    adjusted_weights = []
-
-    for tech, weight in zip(employee_data["Name"], weights):
-        # Consider the historical number of assignments for each tech
-        num_assignments = history[tech]["num_assignments"]
-
-        # Modify the weight based on historical data (customize this part)
-        adjusted_weight = weight * (1.0 + 0.1 * num_assignments)  # Example adjustment
-
-        adjusted_weights.append(adjusted_weight)
-
-    return adjusted_weights
-
-
-
-def update_assignment_data(assignment_data, paired_employees, end_date, start_date):
-    """Update assignment data for each employee."""
-    for tech in paired_employees:
-        assignment_data[tech]["num_assignments"] += 1
-        assignment_data[tech]["last_assignment_date"] = end_date.strftime("%m-%d-%Y")
-        assignment_data[tech]["workload_history"].append(
-            start_date.strftime("%m-%d-%Y")
-        )
-
-
-def write_to_excel(schedule):
-    """Write the schedule to an Excel file with the specified format."""
-    flat_schedule = []
-
-    for week, data in schedule.items():
-        start_date = datetime.strptime(data["start_date"], "%m-%d-%Y")
-        end_date = datetime.strptime(data["end_date"], "%m-%d-%Y")
-
-        flat_schedule.append(
-            {
-                "start_date": start_date.strftime("%m-%d-%Y"),
-                "end_date": end_date.strftime("%m-%d-%Y"),
-                "Agent 1": data["pair"][0],
-                "Email1": data["email_addresses"][0],
-                "Agent 2": data["pair"][1],
-                "Email2": data["email_addresses"][1],
-            }
-        )
-
-    df = pd.DataFrame(flat_schedule)
-
-    # Specify the path within the "working" directory
-    working_directory = "working"
-    output_file_path = os.path.join(working_directory, "assignments.xlsx")
-
-    # Write to Excel file with the specified columns
-    df.to_excel(output_file_path, index=False)
-
-
 def main():
-    global employee_data
-    employee_data = read_employee_data("team_list.xlsx")
+    team_list_path = os.path.join("working", "team_list.xlsx")
+    # assignments_path = os.path.join("working", "assignments.xlsx")
+
+    employee_data = read_employee_data(team_list_path)
 
     if employee_data is not None:
+        log_activity("Backing up existing assignments.")
         backup_existing_assignments()
 
-        # Initialize the assignment data log file
         initialize_assignment_data_log()
 
-        rotation_schedule = generate_rotation_schedule(employee_data, weeks_in_year=52)
+        rotation_schedule = generate_rotation_schedule(employee_data, WEEKS_IN_YEAR)
 
-        # For production use, write the schedule to an Excel file
+        log_assignment_data(rotation_schedule)
         write_to_excel(rotation_schedule)
 
     else:
         logging.error("Exiting program due to errors.")
 
-    # Log the end of the script execution
-    logging.info("Script execution completed.\n")
-
+    log_activity("Script execution completed.\n")
 
 if __name__ == "__main__":
     main()
